[Ada] Fix incorrect binding to MapViewOfFile in s-win32.ads
Despite the "dw" prefix on the name, the dwNumberOfBytesToMap argument
to MapViewOfFile was changed from DWORD to SIZE_T when 64bit Windows
came about.

This change adjusts the binding we have for it in System.Win32
accordingly. For consistency with established practice, an s-win32
specific version of size_t is introduced and g-sercom__mingw.adb is
adjusted to disambiguate between this new size_t and the one already
exposed in System.CRTL.

2019-08-13  Olivier Hainque  <hainque@adacore.com>

gcc/ada/

	* libgnat/s-win32.ads: Define size_t and fix the MapViewOfFile
	binding to use it instead of DWORD for the dwNumberOfBytesToMap
	argument.
	* libgnat/g-sercom__mingw.adb (Read): State which definition of
	size_t to fetch in call to Last_Index.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@274341 138bc75d-0d04-0410-961f-82ee72b054a4